Refactor AddEmployeeForm to improve performance and readability
